Update patch set 4
Patch Set 4:

Still, lgtm.

An exchange with Hans, a bug report, and closer examination of the ARM arch reference manual for ARMv8 makes me think it's not the IC ICAU. 

This appears to deviate from the AArch64 Reference Manual (D5.6.1 types of MMU fault) which says "Cache maintenance instructions cannot cause a Permission fault, except that..." and then states 2 conditions that can raise a permission fault neither of which should apply here if understood correctly. The first of these would be a fault when fetching the data for the TLB. The second specifically relates to DC IVAC. 

Looking at a bug report (https://b.corp.google.com/issues/29434748) it points to line 54 of https://android.googlesource.com/toolchain/gcc/+/master/gcc-4.9/libgcc/config/aarch64/sync-cache.c which is the DC CAU instruction. Could we be in errata or design customization territory here?

Patch-set: 4