Check debuggable flag for quick deoptimized frames
When a quick (optimized) frame is deoptimized into an interpreter
(shadow) frame, it must contain stack map entries for all live virtual
registers. They must be valid for the dex code to be executed further
by the interpreter. This stack map correctness is guaranteed either by
the debuggable flag or HDeoptimize instruction. The latter is applicable
only for a single frame deoptimization.

This patch adds a check that for a multi-frame deoptimization each
frame corresponds to an oat method that was compiled as debuggable.

This patch invalidates several ART tests that deoptimize boot/core oat
frames because they are compiled as non-debuggable.

Test: m test-art-host
Change-Id: I901f991f14a05192249cd37b2d4e5c23dd6acc78
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>