Update patch set 4
Patch Set 4:

> However I think there a crash could occur if
 > - the destination is a stack slot and a source is a constant
 > (different from 0)
 > - and we are emitting code for a move while the parallel move
 > resolve already acquired a scratch core register
 > - and the stack offset at which we are storing cannot be encoded in
 > the STR instruction.

If I understand correctly, the case where the "parallel move resolve already acquired a scratch core register" is for breaking a cycle and we shall only emit moves for that cycle before releasing the scratch register. Since constants never contribute to a cycle, this should not be a problem. Right?

Patch-set: 4