Update patch set 3
Patch Set 3:

I'm not super happy with this change. There are other threads that aren't attached (everything in any ThreadPool). I think this change isn't defensive enough and can potentially mask a breakage there.

(I can't think of a good way of how to fix this, except attaching the JIT thread, or adding a parameter to ResolveType... :-( )

Patch-set: 3