Update patch set 3
Patch Set 3:

> > I'm not super happy with this change. There are other threads
 > that
 > > aren't attached (everything in any ThreadPool). I think this
 > change
 > > isn't defensive enough and can potentially mask a breakage there.
 > >
 > > (I can't think of a good way of how to fix this, except attaching
 > > the JIT thread, or adding a parameter to ResolveType... :-( )
 > 
 > Adding a parameter sounds good to me too.

I was trying to avoid adding an extra parameter. I agree it's probably the easiest solution to reason about but it will be rather messy due to lots of indirection...

Just to clarify. I also think this solution is hard to analyze in all the cases. But so far I couldn't find a counter example when it will not work. I was hopping that maybe Andreas/Mathieu have more context in this area :).

We could submit this while working on the parameter passing if this makes others nervous about possible missed cases.

Patch-set: 3