Update patch set 1
Patch Set 1:

(2 comments)

The idea behind the patch was to be able to fuzz new Vixl based arm backend against the current arm backend. This is why I want to add a dex2oat option. Currently we can only switch backends at build time which is not really convenient for running both of the backends on one device.

Patch-set: 1