Update patch set 2
Patch Set 2:

> Nicolas asked me to do the escape split out stuff in another CL.
 > How else can I get that change rebased into this CL (so that is
 > actually becomes smaller)?

When splitting, you should split at the same parent:

  Before: P -> A#PS1
  After: P -> B -> A#PS2

Then you can do your modifications and submit B, then rebase A and upload A#PS3, then do further modifications on A if needed. That way, the diff A#PS1 A#PS2 is empty and the rebase of A#PS2->A#PS3 is marked as a rebase by gerrit, so I don't need to look at those changes.

Patch-set: 2