Update patch set 3
Patch Set 3:

We were unable to auto-submit your CL due to the following error: Failed to submit 2 changes due to the following problems:

Change 309559: Change could not be merged due to a path conflict. Please rebase the change locally and upload the rebased commit for review.

Change 309560: Change could not be merged due to a path conflict. Please rebase the change locally and upload the rebased commit for review.

Patch-set: 3
Tag: autogenerated:TreeHugger:Autosubmit-Fail