Introduce pre-emptive update of object fields to avoid read barriers.
Introduce a new GC related compiler optimization pass
identifying successive InstanceFieldGet instructions on the
same object and inserting an instruction updating the
corresponding fields in the holder if it is gray
(UpdateFields), thus removing the need for emitting a read
barrier in the InstanceFieldGet instructions.

Note: This CL is a first version to test the idea of
introducing and using HUpdateFields instructions. It could
be extended to cover more InstanceFieldGet instructions (as
well as other types of instructions) and recognize less
rigid patterns.

Test: Run ART tests in various read barrier/heap poisoning configurations.
Test: Boot a device in various read barrier/heap poisoning configurations.
Bug: 29516974
Change-Id: I2c1f1cfaf209a5f620b56250af9c827d9d9b22db