Update patch set 7
Patch Set 7:

(2 comments)

Is this mostly accessed by one thread? Or are writes relatively common?

If neither, I would still try to dodge this issue by using seqlocks instead, i.e. maintain a version count that is incremented before and after each write. Readers read the two halves separately, and check that the version count is even and the same before and after the read. That has the advantage that reads don't modify memory or cache lines.

Patch-set: 7