Update patch set 2
Patch Set 2:

I never understand which things should go into the TLS and which should not. If you don't put it into the TLS, the change will be a lot smaller.

bdc prefers to *not* annotate with things like /*can_call_java*/, which admittedly is annoying when you want to comment out things. He strongly prefers enums now.

Patch-set: 2