Update patch set 1
Patch Set 1:

> (1 comment)

Ah, while constructing the regression test, I know also understand what Nicolas was trying to say. We could actually still proceed escape analysis when the buffer is passed to a string constructor, since that only copies contents. That is an interesting extension, but not one I do in this CL.....

Patch-set: 1