Update patch set 1
Patch Set 1:

> Why isn't vixl capable of doing this?

If current ART backend emits a forward branch it optimistically uses 16bit encoding. If afterwards the range requirement for the bound label is not satisfied, the 32bit version will "squeeze" into the middle of the buffer. Currently VIXL can't do this; it was a design solution inherited from VIXL64.

Patch-set: 1
Label: Verified=0