Update patch set 4
Patch Set 4:

I think this is fine as it is. Since VIXL is managing the register pool generating a constant in the middle of the cycle will mean that VIXL will fire an assert. Also with 3x avaialble registers IP0, IP1, D31 I think none of this is a concern.

Patch-set: 4