Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Not sure how advanced this is, and how committed you are, but please sync with sehr@ and shubhamajmera@, as Shubham has also started working on a rewrite of run-test.

Thanks for the heads up. I was unaware that anyone else was actively working on this. I started making this during the couple of days when nobody was here since it was really hard to figure out what flags a run-test invoke used (also ART_TEST_FULL is totally useless).

Currently this seems to basically work when invoked directly (there might be a few flags that need to be added, but the important ones are there). I haven't made the .mk changes needed to have this be used yet but that shouldn't be too difficult. I'd mostly planed to continue working on this in my spare time but I would like to see this (or some similar improvement) in at some point.

Patch-set: 3
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0