Don't need to block in JavaVMExt::AddWeakGlobalRef under CC.
CMS needs this to block because an object allocated during the GC
won't be marked and concurrent reference processing would incorrectly
clear the JNI weak ref.

But CC doesn't because of the to-space invariant, that is, when a
mutator tries to create a JNI weak ref for an object, it must be
already marked and the concurrent reference processing wouldn't
incorrectly clear it.

Bug: 34128900
Bug: 12687968
Test: test-art-host with CC.
Change-Id: Ia87bf8ed9e604900df5ecb450c89b0ac222bef32