Update patch set 5
Patch Set 5:

PS6 removes the new -XX:NativeAllocationGcWatermark and -XX:NativeAllocationBlockingGcWatermark nobs in favor of deriving these values internally based on max_free_ and growth_limit_ parameters. Because it's hard to believe a user would be able to pick good values for the native watermarks without a fairly deep understanding of the implementation, and max_free_ and growth_limit_ are analogous enough that it makes sense to piggy back on them for native.

Patch-set: 5
Label: Verified=0