Reduce using ArtMethod's dex_cache_resolved_types_.
Avoid using the ArtMethod's dex cache type array shortcut
in runtime, preparing for its removal. We do not completely
remove the shortcut yet because it is still used by array
allocation entrypoints.

Fix ArgArray::BuildArgArrayFromObjectArray in reflection.cc
to not ask for the parameter type to be resolved. It should
have been previously resolved when retrieving the Method.

Fix ResolveCatchBlockExceptionsClassVisitor in
compiler_driver.cc to avoid checking if classes are resolved
while holding the classlinker_classes_lock_. Once we move to
a hashed-based dex cache type array, checking if a class is
resolved shall require a class lookup that takes that lock.

Test: m test-art-host
Bug: 30627598
Change-Id: Ic45a5ff8c66b79429e440cbc08d67bf22a083682