Update patch set 2
Patch Set 2:

(8 comments)

> > (I wonder whether this should go into frameworks/base/cmds. It
 > > isn't really relevant for ART directly, and more a framework
 > > helper. Opinions?)
 > 
 > I think it's easier to keep in sync we potential changes in
 > OatFileAssistant APIs if we keep it in ART. Also, we have the gtest
 > and it's easier to verify they return the same thing.

Given the implementation's heavy dependence on ART, I think having it in art makes sense. As an alternative, if you remove the runtime dependency, it may be natural to implement a new library for dex opt analysis. The library could live in art, and the wrapper command could live in frameworks/base/cmds.

Patch-set: 2
Label: Verified=0