WIP: Also use the holder's gray bit in Baker read barrier slow paths.
In compiler-generated code, when deciding whether to mark
a heap reference or not in a read barrier, after checking
whether the GC is currently marking, also check (in the
slow path) whether the lock word of the reference's holder
is gray, before actually marking the reference.

TODO: Implement for ARM, ARM VIXL, x86 and x86-64.

Test: Run ART tests in Baker read barrier configuration.
Test: Boot a device in Baker read barrier configuration.
Bug: 29516974
Change-Id: I6ee3a4464197a78835f8f01c8a554e46e4b82c02