Improve aarch64 MonitorEntry/Exit assembly code
We make two kinds of changes:

1) We remove some redundant moves, which appeared to have been copied
from some architecture with a 2 address instruction format.

2) We avoid the use of dmb barrier instructions, and instead use
acquire/release instructions for the actual lock loads/updates.

(2) is probably a minor performance loss on some 2016 phones with
Qualcomm cores. But based on microbenchmarking, it seems to be a
win on fully ARM designed cores. On A53/A57, there
seems to be very little additional cost associated with acquire/release
when used with "exclusive" memory operations, as they are here.
ARM's advice for the future is also to move in this direction.

Test: AOSP boots. AOSP art test failures seem attributable to other
issues.

Change-Id: I2399baeab3df93196471e65612c00d95ad4e2b62