Update patch set 4
Patch Set 4:

https://android.googlesource.com/platform/art/+/master/test/Android.run-test.mk#679
  ifeq ($(ART_USE_READ_BARRIER),true)
  […]
  endif

I think this one is fine as is because ART_USE_READ_BARRIER will be true by default.

https://android.googlesource.com/platform/art/+/master/dex2oat/dex2oat.cc#396
  UsageError("  --force-determinism: force the compiler to emit a deterministic output.");
  UsageError("      This option is incompatible with read barriers (e.g., if dex2oat has been");
  UsageError("      built with the environment variable `ART_USE_READ_BARRIER` set to `true`).");
  UsageError("");

I'll defer this to b/34110999.

https://android.googlesource.com/platform/build/+/master/core/soong_config.mk#67
  echo '    "ArtUseReadBarrier": $(if $(filter true,$(PRODUCT_ART_USE_READ_BARRIER)),true,false),'; \

Yes, eventually. That would be another way to enable CC but for internal trees only.  This one can be deferred until enabling CC (this CL) sticks and if we need a way to turn off CC selectively.

Patch-set: 4
Label: Verified=0