Update patch set 2
Patch Set 2:

(2 comments)

Given it's a race, I agree that's going to be hard to test, and the build failing is an indication we'll eventually catch it.

But maybe you could add a stress class loading test, instantiating over and over a class loader that loads classes leading to the situation that broke?

Patch-set: 2