Update patch set 7
Patch Set 7:

We are working to get updated toolchains into the build today, so perhaps this will be an already fixed issue. Since all of my branches are tied up with validation for that today, if you can get me the preprocessed source + command line for the crash, I can try with the latest binaries. If not, you will have to wait to resubmit this part after the upgrade is complete (and then we can investigate as necessary).

Patch-set: 7
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0