Update patch set 1
Patch Set 1:

> (1 comment)
As Scott has described to have a test covering this issue we need to have either veneer or literal pools with ~1024 entries emitted between ADR and table itself to exceed the ADR range.
Currently the only ways to do this (and of course forming a specific pattern):
1. (veneer pool) Have a method size ~ 1Mb to trigger veneer emission for T3 encoding of B.cond.
2. (literal pool) Emit ~1024 HInvokeSt.. to different methods or ~1024 HLoadClass of different classes with dispatch kind which uses literals.

Do you think it is reasonable to do now? Later, when we introduce literals for double constants it will be easier to form a regression test.

Patch-set: 1
Reviewer: Gerrit User 1089808 <1089808@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0