Update patch set 1
Patch Set 1:

Thanks Roland for the fix!

Serban, Scott: this is again a problem with the NoSwap ParallelMove version. Roland's fix makes things less likely to break, but it can still happen. I was under the impression the NoSwap version was *guaranteeing* that we would not run out of registers (hence no need to ever swap).

So should we remove the non-functional NoSwap version?

Patch-set: 1
Label: Verified=0