Update patch set 3
Patch Set 3: Code-Review+1

> Thanks Roland for the fix!
 > 
 > Serban, Scott: this is again a problem with the NoSwap ParallelMove
 > version. Roland's fix makes things less likely to break, but it can
 > still happen. I was under the impression the NoSwap version was
 > *guaranteeing* that we would not run out of registers (hence no
 > need to ever swap).
 > 
 > So should we remove the non-functional NoSwap version?

ParallelMoveResolverNoSwap was written some time ago and was not touched afterwards while VIXL has moved. So now we've opened a ticket to reexamine it and check that it never exhausts scratch registers; also will extend unit tests for the resolver if something is not covered by them.

This particular fix seems to be fine.

Patch-set: 3
Label: Code-Review=+1