Fix flaky jit-gcstress tests 93{4,5,8}.
We were letting the jit and the main thread race on class definition.
If the 'leader' changes then we would end up with the test class
having an unexpected definition, causing the test to fail. We fixed
this by making sure all threads get the same definition regardless of
speed.

Bug: 34799243
Test: mma -j40 test-art-host
Change-Id: I0d6785b4c3c41db1d655cf1ff22be60809923a7a