build: Generate asm_support_gen.h from the build, validate up-to-date
When building libart/libartd, also generate asm_support_gen.h
automatically. In addition, verify that our checked-in version
(runtime/generated/asm_support_gen.h) is up-to-date with what the build
generates.

This makes it significantly more difficult to accidentally
merge a CL that doesn't have the auto-generated headers updated
and break the build as a result.

Bug: 34387670
Change-Id: I1f0f94914d328c396906583d0732e281c076e69f