Update patch set 1
Patch Set 1:

I think, as Mathieu says, a stale ref could be a problem and the old field value is likely null.

How about not calling ModifyDebugDisallowReadBarrier in ConcurrentCopying::Scan if in transaction mode? It was for detecting (and avoiding) unexpected read barriers during GC scanning for a *performance* reason. It is probably okay to not to do this in transaction mode.

Patch-set: 1
Reviewer: Gerrit User 1022530 <1022530@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0