Update patch set 1
Patch Set 1:

> > > Thanks, you did well to reproduce that. :)
 > > >
 > > > We will discuss this with the VIXL team soon.
 > >
 > > Thanks!
 > 
 > The reason for the bug is incorrect usage of BlockPoolsScope in
 > code_generator_arm64.cc; it is strange that regression has happened
 > only now. I am currently working on a fix.

Cool.  I was indeed thinking that the breakage was due to a wrong use of a BlockPoolsScope.  I guess we're only seeing this now as this is an edge case: the original code that revealed the crash was generated by a tool and contains an extremely long switch.

Patch-set: 1