Update patch set 2
Patch Set 2:

PTAL.  I want a preliminary check that I'm headed in the right direction.  I will write the test once that's done.  I envision the test being two runs of profman, and a grep | sort | uniq to make sure the class list is the same as for input.  Other suggestions are welcome, of course.

Patch-set: 2
Label: Verified=0