Update patch set 1
Patch Set 1:

Opinions on this? This would allow moving the 9:30 over to installd, instead of us having to have knowledge here (plus allow to make it longer if we get around to needing that).

If you're OK with it, I'll add a dex2oat_test and remove WIP.

Patch-set: 1
Label: Presubmit-Ready=0