Update patch set 1
Patch Set 1:

Seems that my laziness of not putting the obsolete methods into the classes obsolete methods array immediately is coming back to bite me.

When iterating though a stack with native methods in it we sometimes need to get the methods dex cache. Unfortunately we have set the method to be the obsolete method but haven't filled in the classes map yet. This means that we aren't able to find the dex cache causing an assertion failure.

This shouldn't be too difficult to fix hopefully but it will include some reworking of the redefinition code.

Patch-set: 1
Label: Verified=0