Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review-2
> 
> (-2 as the fix can be done without looking at obsolete methods.
> 
> I believe the problem is that ArtMethod::GetOatQuickMethodHeader leads to FindOatMethod for native methods. But that can actually be avoided given that a native method entrypoint won't change, and you can just do:
> 
> (do this after generic jni check and friends)
> 
> if (method->IsNative()) {
>   // We know the entrypoint in the ArtMethod is always the same.
>   return OatQuickMethodHeader::FromEntryPoint(existing_entry_point);
> }

I'm not sure we can from experimentation. Also the underlying problem is that not all methods of artMethod could be called during a brief time between when the obsolete method is created and it being fully installed. During this time it might not be possible for an obsolete method to find it's own dex file. Fixing this issue fixes (almost) all the issues we were having with the stack-walking too.

Patch-set: 2
Label: Verified=0