Testrunner: Differentiate between flag set to False and not set.
When flags ART_TEST_OPTIMIZING, ART_TEST_RUN_TEST_PREBUILD,
ART_TEST_RUN_TEST_IMAGE and ART_TEST_RUN_TEST_DEBUG were not set,
the testrunner was assuming them to be True. As a result,
even though the tests are not asked for, the testrunner used to
run the tests assuming that they have been asked through environmental
variables.

Now, if the flags are not set, instead of True, they will be
marked as None. By this, it can be find out whether they have been
explicitly requested for or not.

So, if the value is unset, the testrunner will only used the variant
if no other variant of that type has been requested through command
line arguments.

Test: Checked the behavior by running locally.
Change-Id: I6f480249408c0bf406b660e5098f298f21ff2db4