Update patch set 1
Patch Set 1:

As Nicolas and I discussed, we decided to keep one dependency set for run-test for host and one for target. This makes easier to build dependencies. Otherwise, we will have to keep dependencies on environment variables like ART_TEST_INTERPRETER.
Anyway, this is a temporary fix to keep the builds green. We can discuss it further maybe?

Patch-set: 1