Update patch set 3
Patch Set 3:

I can see how PS2 could be considered a right approach in some situations.

The 071-dexfile-map-clean test is meant to check that dex files are mapped as clean, but that's obviously impossible if it's on a TMPFS. That could be construed as bad test *infrastructure* because it should somehow be running with DEX_LOCATION not in a tmpfs.

Running the test against DEX_LOCATION in tmpfs is pointless because it is a no-op (either vacuously true with a check script or vacuously false without one). That is why most testing frameworks have a way to mark tests as skipped when preconditions fail, rather than passing it.

Consider if in the future we ever switch our test/build servers to use TMPFS, it will gratuitously report 'PASSED' and we will never spot any regressions.

From that respect I believe SKIPping the test is the most correct approach. Since we don't have a way in run-test to mark tests as skipped (and that would require a larger change), this seems like a fine and very simple alternative.

Patch-set: 3