Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Shubham, we had this functionality in our Makefiles previously. I seem to recall a lot of people to use it to skip the very long running tests.
> 
> The new testrunner appears to have lost this during the rewrite.

I am not exactly disagreeing with having the flag support, but at the same time we can also add an option for skipping like --skip. I would prefer of having an equivalent cmd line option for env var.
One point supporting '--skip' option would be - if you are running testrunner with multiple configuration, and want to disable a test for a particular run, it would be convenient to just pass it as an option rather than setting env var, and unsetting it for other runs with which we don't want the test to be skipped.

Patch-set: 3
Label: Verified=0