Remove String.setCharAt().
The internal API String.setCharAt() breaks the assumption
that strings are really immutable. That in turn breaks
string compression invariants - compressible strings must
be compressed. This removes the

The method was used only in String.replace(char, char)
when we found a match, copying the string on first match.
Instead, introduce a new native method that does the whole
replacement with a single call when we find the first match.

TODO: Run StringReplacementBenchmark.

Test: testrunner.py --host
Test: run-libcore-tests.sh --mode=host
Bug: 31040547
Change-Id: I9cf0d5457182f0a33ca8251c29931d3eb624ae07