Update patch set 1
Patch Set 1:

> Is there a reason that this shouldn't be combined with the the
 > IsGcMarking strategy CL?

This CL conflicts with the other CLs (324303, 341285, 334544) that are based on the « Use "Is Gc Marking" info » base CL (321519).  This is why I have kept these separated for now, so as to make amending and rebasing easier.

Eventually (after I get updated benchmark numbers), the idea is to land a pair of this CL: 321519 + one of the others (and to abandon the other ones).

Patch-set: 1
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>