Dexlayout fix for duplicate class data and preserving last code item.
This change makes dexlayout preserve the offset of the last code item in
the dex file. This is done because code items are 4-byte aligned, but
the section following the code items need not be. The new layout may
need extra space, and dexlayout has no good way to move sections.
An overhaul is needed to handle layout of all sections of the file.

Bug: 35451910
Test: mm test-art-host-gtest-dexlayout_test
Change-Id: I8d066c2151a1a57c382ce35f12bf53a519da89f6