Update patch set 1
Patch Set 1:

Can this be merged as is, or the patch in platform/build (https://android-review.googlesource.com/#/c/327858/)? Both mips32 and mips64 can't boot for several weeks now. I know that you plan to remove product option soon, but we are working on implementation of read barriers for mips and hopefully we will finish it before you remove product option.

As I said in the previous comment, it seems to me that we would need to change a lot of stuff in ART to ignore ART_USE_READ_BARRIERS when it is not supported.

Patch-set: 1
Reviewer: Gerrit User 1067943 <1067943@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0