Don't send ClassFileLoadHook before JVMTI_PHASE_START.
All the classes we load prior to transitioning into JVMTI_PHASE_START
are tightly coupled into the runtime itself. For this reason
redefining them in the ways allowed by the ClassFileLoadHook during
first load could cause difficult to diagnose problems with the
runtime. Since there is no good way to determine if the new definition
is safe nor to report that the definition was ignored we instead
simply do not send this event during this stage of runtime
initialization.

Bug: 35803394
Test: ./test/testrunner/testrunner.py --host -j40 -b
Change-Id: I813f2ced49a2a85e485a6f2cf94edb024799c5c0