DO NOT MERGE: WIP: tools: Add golem/build-target script
Refactor dart logic to build golem .tar.gz into art.

The intent here is to move the logic of building the ART binaries
back into ART directory. This makes it easier to keep the two in sync,
and makes it easier for us to change it later to run from a
regular master build instead of a special golem-only master-art build.

Bug: 35766356
Change-Id: I9c9d7f522da3f850bc841d4c1f6e42c3055b8746