tools: Add golem/build-target and golem/env scripts.
Refactor the golem build logic back into the art directory.

Usage example:

  # set environment variables to build art-opt-cc
  # creates android-armv8.tar.gz, which contains
  # 'dalvikvm' and its libraries that can be used in
  # golem
  art/tools/golem/build-target.sh -j32 --showcommands \
     --config=android-armv8 --golem=art-opt-cc \
     --tarball=android-armv8.tar.gz
  # (drop the --golem to build with your own lunch combo)

The intent here is to move the logic of building the ART binaries
back into ART directory. This makes it easier to keep the two in sync,
and makes it easier for us to change it later to run from a
regular master build instead of a special golem-only master-art build.

Bug: 35766356
Change-Id: I9c9d7f522da3f850bc841d4c1f6e42c3055b8746