Update patch set 2
Patch Set 2: Code-Review-1

(4 comments)

This has really reduced the readability of the testrunner and at one point it was difficult for me to follow up.

IMO. in the existing implementation, semaphore are not interfering with the testrunner logic as much as they are doing now.

IMO, it would become difficult to manage new changes in the testrunner with this implementation.

Patch-set: 2
Reviewer: Gerrit User 1084268 <1084268@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1