Update patch set 1
Patch Set 1:

Is this only an issue for compilation with GCC or does this manifest with Clang as well? I'm not sure that there is a clean way to change libc++ to address this (or if they would even be open to changing this). I will point out my bug about unintentional use of -fstrict-aliasing in Android (https://b.corp.google.com/issues/31600139).

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>