Update patch set 1
Patch Set 1:

(1 comment)

> I'm not sure what the makefile sets OUT_DIR.

Primarily it's an argument to the build system, and internally the build system defaults it to "out" (unlike this script, which was defaulting it to "out/").

I recently changed the build system to always evaluate OUT_DIR as part of a change to ensure that OUT_DIR_COMMON_BASE was always respected as well, which would change the default of OUT_DIR to $OUT_DIR_COMMON_BASE/$(basename $PWD).
 
> However we do have the option of running test/run-test directly from command line and in those cases it does not set OUT_DIR.
> 
>   $> source build/envsetup.sh
>   $> lunch aosp_angler-userdebug
>   $> echo $OUT_DIR
> 
> results in a blank line being printed

And that's fine, that's what the :-$(ANDROID_BUILD_TOP)/out default is for in run-test.

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>